deployment/requirements.txt
scikit-learn == 1.6.1
pandas
matplotlib
seaborn
numpy
pillow